+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+
 ## 11.1.0
 
 ### Added
@@ -42,6 +43,12 @@
     registration of a given audit record.
   - `VerificationAuditRecord.createChecked` which creates a verification audit record while verifying the presentation in
     the context of the verification request.
+
+- deserialize function for DeployModulePayload implemented instead of throwing an exception of not supporting deserialize
+
+- deserialize function for InitContractPayload implemented instead of throwing an exception of not supporting deserialize
+
+- deserialize function for UpdateContractPayload implemented instead of throwing an exception of not supporting deserialize
 
 ## 11.0.0